Added new property to SFan Performance Class for if a VFD is Present. When Importing a UPG unit, the VFD Present is automatically selected if appropriate.
LCVAV Module:  if ZR is in model number then autoselect HS Relay.
LCVAV Module:  If SFan VFD Present, auto select in this module.
LCVAV Module:  Added a WAG weight for the Series 100 Unit.  I doubt that we would ever need to do this, but it's in place should we need to.
LCVAV Module:  Added core code for Design Cautions.